feat: Split notification code
